TST: Move main conftest.py to root level because no astropy test runner. Handle upstream warning that appears because of this.
